Renamed PlaylistService to DefaultPlaylistService and made the DefaultPlaylistService implement interface
