+++ src/main/java/com/odeyalo/sonata/playlists/config/PlaylistItemsConfiguration.java
@@ -1,8 +1,5 @@
 package com.odeyalo.sonata.playlists.config;
 
-import com.odeyalo.sonata.playlists.repository.PlaylistRepository;
-import com.odeyalo.sonata.playlists.service.PlaylistLoader;
-import com.odeyalo.sonata.playlists.service.PlaylistService;
 import com.odeyalo.sonata.playlists.service.tracks.InMemoryPlayableItemLoader;
 import com.odeyalo.sonata.playlists.service.tracks.PlayableItemLoader;
 import org.springframework.context.annotation.Bean;
@@ -14,10 +11,5 @@ public class PlaylistItemsConfiguration {
     @Bean
     public PlayableItemLoader playableItemLoader() {
         return new InMemoryPlayableItemLoader();
-    }
-
-    @Bean
-    public PlaylistLoader playlistLoader(PlaylistRepository playlistRepository) {
-        return new PlaylistService(playlistRepository);
     }
 }
