+++ src/test/java/testing/factory/PlaylistLoaders.java
@@ -3,7 +3,7 @@ package testing.factory;
 import com.odeyalo.sonata.playlists.model.Playlist;
 import com.odeyalo.sonata.playlists.repository.InMemoryPlaylistRepository;
 import com.odeyalo.sonata.playlists.service.PlaylistLoader;
-import com.odeyalo.sonata.playlists.service.PlaylistService;
+import com.odeyalo.sonata.playlists.service.DefaultPlaylistService;
 
 /**
  * Factory methods for {@link PlaylistLoader}
@@ -15,7 +15,7 @@ public final class PlaylistLoaders {
     }
 
     public static PlaylistLoader withPlaylists(Playlist... playlists) {
-        return new PlaylistService(
+        return new DefaultPlaylistService(
                 new InMemoryPlaylistRepository(playlists)
         );
     }
