+++ src/test/java/testing/factory/PlaylistOperationsTestableFactory.java
@@ -4,7 +4,7 @@ import com.odeyalo.sonata.playlists.model.Playlist;
 import com.odeyalo.sonata.playlists.repository.InMemoryPlaylistRepository;
 import com.odeyalo.sonata.playlists.service.DefaultPlaylistOperations;
 import com.odeyalo.sonata.playlists.service.PlaylistOperations;
-import com.odeyalo.sonata.playlists.service.PlaylistService;
+import com.odeyalo.sonata.playlists.service.DefaultPlaylistService;
 import com.odeyalo.sonata.playlists.service.upload.MockImageUploader;
 
 import java.util.List;
@@ -13,11 +13,11 @@ public class PlaylistOperationsTestableFactory {
 
     public static PlaylistOperations withPlaylists(List<Playlist> playlists) {
         final InMemoryPlaylistRepository repository = new InMemoryPlaylistRepository(playlists);
-        return new DefaultPlaylistOperations(new PlaylistService(repository), new MockImageUploader());
+        return new DefaultPlaylistOperations(new DefaultPlaylistService(repository), new MockImageUploader());
     }
 
     public static DefaultPlaylistOperations create() {
         final InMemoryPlaylistRepository repository = new InMemoryPlaylistRepository();
-        return new DefaultPlaylistOperations(new PlaylistService(repository), new MockImageUploader());
+        return new DefaultPlaylistOperations(new DefaultPlaylistService(repository), new MockImageUploader());
     }
 }
